requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
bcrypt==4.1.2
google-generativeai==0.5.4
python-dotenv==1.0.1
tzdata==2024.1
gunicorn
